last fixes (doc + test + code example => src)
