+++ index.html
@@ -100,6 +100,7 @@
 
       createContainer('AllWidget');
       createContainer('SinglePlayerGamePlanar');
+      createContainer('MultiPlayerGamePlanar');
       createContainer('Frame3D');
 
       const addProjectElement = (name, idParent, href, title, srcImg) => {
@@ -262,6 +263,14 @@
         'SinglePlayerGamePlanar',
         './packages/browser/examples/DragAndDropAvatar.html',
         'This example allow to drag and drop an avatar on city',
+        './packages/browser/examples/assets/img/mosaic/AvatarGame.png'
+      );
+
+      addProjectElement(
+        'Note Game',
+        'MultiPlayerGamePlanar',
+        './packages/browser/examples/MultiPlayerGamePlanar.html',
+        'This example allow to share note in city',
         './packages/browser/examples/assets/img/mosaic/AvatarGame.png'
       );
     </script>
