+++ packages/browser/examples/MultiPlayerGamePlanar.html
@@ -7,8 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
   <body>
-    <!-- <script src="../dist/release/bundle.js"></script> -->
-    <script src="../dist/debug/bundle.js"></script>
+    <script src="../dist/RUN_MODE/bundle.js"></script>
 
     <script>
       udvizBrowser.FileUtil.loadMultipleJSON([
@@ -56,7 +55,8 @@
               new udvizBrowser.InputManager(),
               {
                 externalGameScriptClass: {
-                  CameraManager: udvizBrowser.ExternalScriptTemplate.CameraManager,
+                  CameraManager:
+                    udvizBrowser.ExternalScriptTemplate.CameraManager,
                   NoteService: udvizBrowser.ExternalScriptTemplate.NoteService,
                   NoteUI: udvizBrowser.ExternalScriptTemplate.NoteUI,
                   Note: udvizBrowser.ExternalScriptTemplate.Note,
