+++ packages/browser/src/Component/SocketIOWrapper.js
@@ -20,15 +20,12 @@ export class SocketIOWrapper {
         ? 'wss'
         : 'ws';
 
-      // const loc = `${window.location.pathname}`;
-      // const path = loc.substring(0, loc.lastIndexOf('/'));
-
       // Instantiate socket and connect to the server serving index.html
       this.socket = io(`${socketProtocol}://${window.location.host}`, {
         reconnection: false,
         secure: true,
         transports: ['polling', 'websocket'],
-        // path: path + '/socket.io/',
+        autoConnect: true,
       });
 
       // connected
