--- packages/node/examples/script/gameThread.js
@@ -1,2 +1,0 @@
-const Game = require('../../src/index').Game;
-Game.Thread.Routine();
