+++ packages/node/package.json
@@ -23,7 +23,9 @@
     "@ud-viz/shared": "3.0.0",
     "express": "^4.17.1",
     "child-process-promise": "^2.2.1",
+    "express": "^4.18.2",
     "puppeteer": "^19.4.1",
+    "socket.io": "^4.5.4",
     "string-replace-middleware": "^1.0.2"
   },
   "devDependencies": {
