+++ packages/node/src/Game/SocketService.js
@@ -5,19 +5,44 @@ const SocketWrapper = require('./SocketWrapper');
 const GameThread = require('./Thread');
 
 module.exports = class SocketService {
+  /**
+   * Websocket game service
+   *
+   * @param {import('http').Server} httpServer - http server
+   * @param {object} options - options
+   * @param {number} [options.pingInterval=2000] - ping interval of the socket connection in ms
+   * @param {number} [options.pingTimeout=5000] - ping timeout in ms
+   */
   constructor(httpServer, options = {}) {
-    const io = socketio(httpServer, {
+    /**
+     * @type {socketio.Server}
+     */
+    this.io = new socketio.Server(httpServer, {
       pingInterval: options.pingInterval || 2000,
       pingTimeout: options.pingTimeout || 5000,
     });
 
-    io.on('connection', this.onSocketConnection.bind(this));
+    this.io.on('connection', this.onSocketConnection.bind(this));
 
     /** @type {Object<string,GameThread>} - threads running a gamecontext */
     this.threads = {};
 
-    /** @type {Object<string,SocketWrapper>} - socket currently connected */
+    /** @type {Object<string,SocketWrapper>} - socket wrappers currently connected */
     this.socketWrappers = {};
+  }
+
+  /**
+   * Stop threads + disconnect socket client + close websocket connection
+   */
+  stop() {
+    for (const key in this.threads) {
+      const thread = this.threads[key];
+      thread.post(Thread.EVENT.STOP, {});
+    }
+
+    this.io.disconnectSockets();
+
+    this.io.close();
   }
 
   /**
@@ -25,20 +50,14 @@ module.exports = class SocketService {
    *
    * @param {string[]} gameScriptsPath - class needed by game context
    * @param {Game.Object3D} gameObjects3D - gameobject3D to simulate
-   * @param {string} threadPath - path to game thread routine
    * @param {string=} entryGameObject3DUUID - uuid of default gameobject to connect socket connected
    */
-  initializeGameThreads(
-    gameScriptsPath,
-    gameObjects3D,
-    threadPath,
-    entryGameObject3DUUID
-  ) {
+  initializeGameThreads(gameScriptsPath, gameObjects3D, entryGameObject3DUUID) {
     // default gameobject3D when socket connect
     this.entryGameObject3DUUID = entryGameObject3DUUID || gameObjects3D[0].uuid;
 
     gameObjects3D.forEach((gameObject3D) => {
-      this.threads[gameObject3D.uuid] = new Thread(threadPath);
+      this.threads[gameObject3D.uuid] = new Thread();
       this.threads[gameObject3D.uuid].post(Thread.EVENT.INIT, {
         gameScriptsPath: gameScriptsPath,
         gameObject3D: gameObject3D,
